Fix duplication by Coporation and CorporateBody. Replace Corporation bei CorporateBody
