+++ dc-model-jackson/src/main/java/de/digitalcollections/model/jackson/DigitalCollectionsModelModule.java
@@ -14,7 +14,6 @@ import de.digitalcollections.model.api.identifiable.IdentifierType;
 import de.digitalcollections.model.api.identifiable.Node;
 import de.digitalcollections.model.api.identifiable.entity.Article;
 import de.digitalcollections.model.api.identifiable.entity.Collection;
-import de.digitalcollections.model.api.identifiable.entity.Corporation;
 import de.digitalcollections.model.api.identifiable.entity.DigitalObject;
 import de.digitalcollections.model.api.identifiable.entity.Entity;
 import de.digitalcollections.model.api.identifiable.entity.EntityRelation;
@@ -80,7 +79,6 @@ import de.digitalcollections.model.jackson.mixin.identifiable.NodeMixIn;
 import de.digitalcollections.model.jackson.mixin.identifiable.VersionMixIn;
 import de.digitalcollections.model.jackson.mixin.identifiable.entity.ArticleMixIn;
 import de.digitalcollections.model.jackson.mixin.identifiable.entity.CollectionMixIn;
-import de.digitalcollections.model.jackson.mixin.identifiable.entity.CorporationMixIn;
 import de.digitalcollections.model.jackson.mixin.identifiable.entity.DigitalObjectMixIn;
 import de.digitalcollections.model.jackson.mixin.identifiable.entity.EntityMixIn;
 import de.digitalcollections.model.jackson.mixin.identifiable.entity.EntityRelationMixIn;
@@ -197,7 +195,6 @@ public class DigitalCollectionsModelModule extends SimpleModule {
     context.setMixInAnnotations(CodeBlock.class, CodeBlockMixIn.class);
     context.setMixInAnnotations(Collection.class, CollectionMixIn.class);
     context.setMixInAnnotations(ContentBlock.class, ContentBlockMixIn.class);
-    context.setMixInAnnotations(Corporation.class, CorporationMixIn.class);
     context.setMixInAnnotations(DigitalObject.class, DigitalObjectMixIn.class);
     context.setMixInAnnotations(Entity.class, EntityMixIn.class);
     context.setMixInAnnotations(EntityRelation.class, EntityRelationMixIn.class);
