--- dc-model-jackson/src/main/java/de/digitalcollections/model/jackson/mixin/identifiable/entity/CorporationMixIn.java
@@ -1,11 +1,0 @@
-package de.digitalcollections.model.jackson.mixin.identifiable.entity;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonTypeName;
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import de.digitalcollections.model.impl.identifiable.entity.CorporationImpl;
-
-@JsonDeserialize(as = CorporationImpl.class)
-@JsonIgnoreProperties(ignoreUnknown = true)
-@JsonTypeName("CORPORATION")
-public interface CorporationMixIn {}
