+++ dc-model-jackson/src/test/java/de/digitalcollections/model/jackson/identifiable/entity/EntityTest.java
@@ -1,8 +1,8 @@
 package de.digitalcollections.model.jackson.identifiable.entity;
 
 import de.digitalcollections.model.api.identifiable.entity.Entity;
-import de.digitalcollections.model.impl.identifiable.entity.CorporationImpl;
 import de.digitalcollections.model.impl.identifiable.entity.ProjectImpl;
+import de.digitalcollections.model.impl.identifiable.entity.agent.CorporateBodyImpl;
 import de.digitalcollections.model.impl.identifiable.parts.LocalizedTextImpl;
 import de.digitalcollections.model.jackson.BaseJsonSerializationTest;
 import java.util.Locale;
@@ -13,11 +13,11 @@ import org.junit.jupiter.api.Test;
 public class EntityTest extends BaseJsonSerializationTest {
 
   @Test
-  @DisplayName("works with a corporation as entity")
+  @DisplayName("works with a corporate body as entity")
   public void testSerialisationForCorporationsInBothWays() throws Exception {
-    Entity corporation = new CorporationImpl();
-    corporation.setLabel(new LocalizedTextImpl(Locale.GERMAN, "Bayerische Staatsbibliothek"));
-    checkSerializeDeserialize(corporation);
+    Entity corporateBody = new CorporateBodyImpl();
+    corporateBody.setLabel(new LocalizedTextImpl(Locale.GERMAN, "Bayerische Staatsbibliothek"));
+    checkSerializeDeserialize(corporateBody);
   }
 
   @Test
