+++ dc-model-xml/src/main/java/de/digitalcollections/model/xml/xstream/DigitalCollectionsXStreamMarshaller.java
@@ -4,10 +4,10 @@ import com.thoughtworks.xstream.XStream;
 import de.digitalcollections.model.api.identifiable.parts.structuredcontent.ContentBlock;
 import de.digitalcollections.model.impl.identifiable.entity.ArticleImpl;
 import de.digitalcollections.model.impl.identifiable.entity.CollectionImpl;
-import de.digitalcollections.model.impl.identifiable.entity.CorporationImpl;
 import de.digitalcollections.model.impl.identifiable.entity.ProjectImpl;
 import de.digitalcollections.model.impl.identifiable.entity.TopicImpl;
 import de.digitalcollections.model.impl.identifiable.entity.WebsiteImpl;
+import de.digitalcollections.model.impl.identifiable.entity.agent.CorporateBodyImpl;
 import de.digitalcollections.model.impl.identifiable.entity.parts.SubtopicImpl;
 import de.digitalcollections.model.impl.identifiable.entity.parts.WebpageImpl;
 import de.digitalcollections.model.impl.identifiable.parts.TranslationImpl;
@@ -79,7 +79,7 @@ public class DigitalCollectionsXStreamMarshaller extends XStreamMarshaller {
     aliases.put("codeBlock", CodeBlockImpl.class);
     aliases.put("collection", CollectionImpl.class);
     aliases.put("contentBlocks", ContentBlock[].class);
-    aliases.put("corporation", CorporationImpl.class);
+    aliases.put("corporateBody", CorporateBodyImpl.class);
     aliases.put("fileresource", FileResourceImpl.class);
     aliases.put("hardBreak", HardBreakImpl.class);
     aliases.put("heading", HeadingImpl.class);
