+++ dc-model-xml/src/main/java/de/digitalcollections/model/xml/xstream/v1/V1DigitalCollectionsXStreamMarshaller.java
@@ -4,9 +4,9 @@ import com.thoughtworks.xstream.XStream;
 import de.digitalcollections.model.api.identifiable.parts.structuredcontent.ContentBlock;
 import de.digitalcollections.model.impl.identifiable.entity.ArticleImpl;
 import de.digitalcollections.model.impl.identifiable.entity.CollectionImpl;
-import de.digitalcollections.model.impl.identifiable.entity.CorporationImpl;
 import de.digitalcollections.model.impl.identifiable.entity.ProjectImpl;
 import de.digitalcollections.model.impl.identifiable.entity.WebsiteImpl;
+import de.digitalcollections.model.impl.identifiable.entity.agent.CorporateBodyImpl;
 import de.digitalcollections.model.impl.identifiable.entity.parts.WebpageImpl;
 import de.digitalcollections.model.impl.identifiable.parts.TranslationImpl;
 import de.digitalcollections.model.impl.identifiable.parts.structuredcontent.StructuredContentImpl;
@@ -80,7 +80,7 @@ public class V1DigitalCollectionsXStreamMarshaller extends XStreamMarshaller {
     aliases.put("collection", CollectionImpl.class);
     aliases.put("codeBlock", CodeBlockImpl.class);
     aliases.put("content", ContentBlock.class);
-    aliases.put("corporation", CorporationImpl.class);
+    aliases.put("corporateBody", CorporateBodyImpl.class);
     aliases.put("fileresource", FileResourceImpl.class);
     aliases.put("hardBreak", HardBreakImpl.class);
     aliases.put("heading", HeadingImpl.class);
